#13 목요일 followup

+++ Jinwoo/Project/Engine/CEngine.cpp
@@ -29,7 +29,7 @@ int CEngine::init(HWND _hWnd, Vec2 _vResolution)
 	RECT rt = { 0, 0, (int)m_vResolution.x, (int)m_vResolution.y };
 
 	AdjustWindowRect(&rt, WS_OVERLAPPEDWINDOW, false);
-	SetWindowPos(m_hMainWnd, nullptr, (int)10.f, (int)10.f, rt.right - rt.left, rt.bottom - rt.top, 0);
+	SetWindowPos(m_hMainWnd, nullptr, 10, 10, rt.right - rt.left, rt.bottom - rt.top, 0);
 
 	if (CDevice::GetInst()->init(m_hMainWnd, m_vResolution))
 	{
